Update sfpegRecordEditFlw.js
for fixing bug - if we try to remove any value from the field then it doesnt include that field in new record list.


Because the value returned is undefined so it doesnt include that field where value is being removed. So if user tries to remove lookup then it will not allow that.

Solution: If its undefined then override it with null this way it will be automatically included in the newRecord List which in turn be returned to output then that can be updated later.